Refactor SHA256 hashing logic into mysql utility module
Moved MySQL SHA256 hashing and random salt generation logic from proxysql_mysql_users.py to module_utils/mysql.py. This improves code reuse, avoids duplication, and centralizes shared utilities for easier maintenance and updates.
