+++ plugins/modules/proxysql_mysql_users.py
@@ -174,139 +174,21 @@ from ansible_collections.community.proxysql.plugins.module_utils.mysql import (
     proxysql_common_argument_spec,
     save_config_to_disk,
     load_config_to_runtime,
+    mysql_sha256_password_hash,
+    generate_random_salt,
 )
 
 from ansible.module_utils.six import iteritems
 from ansible.module_utils._text import to_native, to_bytes
-from hashlib import sha1, sha256
-from os import urandom
-from base64 import urlsafe_b64encode
+from hashlib import sha1
+
 # ===========================================
 # proxysql module specific support methods.
 #
 
-
-# Imported code from @Aohzan
-# community.mysql/plugins/module_utils/implementations/mysql/hash.py
-def _to64(v, n):
-    """Convert a 32-bit integer to a base-64 string"""
-    i64 = (
-        [".", "/"]
-        + [chr(x) for x in range(48, 58)]
-        + [chr(x) for x in range(65, 91)]
-        + [chr(x) for x in range(97, 123)]
-    )
-    result = ""
-    while n > 0:
-        n -= 1
-        result += i64[v & 0x3F]
-        v >>= 6
-    return result
-
-
-def _hashlib_sha256(data):
-    """Return SHA-256 digest from hashlib ."""
-    return sha256(data).digest()
-
-
-def _sha256_digest(key, salt, loops):
-    """Return a SHA-256 digest of the concatenation of the key, the salt, and the key, repeated as necessary."""
-    # https://www.akkadia.org/drepper/SHA-crypt.txt
-    num_bytes = 32
-    bytes_key = key.encode()
-    bytes_salt = salt.encode()
-    digest_b = _hashlib_sha256(bytes_key + bytes_salt + bytes_key)
-
-    tmp = bytes_key + bytes_salt
-    for i in range(len(bytes_key), 0, -num_bytes):
-        tmp += digest_b if i > num_bytes else digest_b[:i]
-
-    i = len(bytes_key)
-    while i > 0:
-        tmp += digest_b if (i & 1) != 0 else bytes_key
-        i >>= 1
-
-    digest_a = _hashlib_sha256(tmp)
-
-    tmp = b""
-    for i in range(len(bytes_key)):
-        tmp += bytes_key
-
-    digest_dp = _hashlib_sha256(tmp)
-
-    byte_sequence_p = b""
-    for i in range(len(bytes_key), 0, -num_bytes):
-        byte_sequence_p += digest_dp if i > num_bytes else digest_dp[:i]
-
-    tmp = b""
-    til = 16 + digest_a[0]
-
-    for i in range(til):
-        tmp += bytes_salt
-
-    digest_ds = _hashlib_sha256(tmp)
-
-    byte_sequence_s = b""
-    for i in range(len(bytes_salt), 0, -num_bytes):
-        byte_sequence_s += digest_ds if i > num_bytes else digest_ds[:i]
-
-    digest_c = digest_a
-
-    for i in range(loops):
-        tmp = byte_sequence_p if (i & 1) else digest_c
-        if i % 3:
-            tmp += byte_sequence_s
-        if i % 7:
-            tmp += byte_sequence_p
-        tmp += digest_c if (i & 1) else byte_sequence_p
-        digest_c = _hashlib_sha256(tmp)
-
-    inc1, inc2, mod, end = (10, 21, 30, 0)
-
-    i = 0
-    tmp = ""
-
-    while True:
-        tmp += _to64(
-            (digest_c[i] << 16)
-            | (digest_c[(i + inc1) % mod] << 8)
-            | digest_c[(i + inc1 * 2) % mod],
-            4,
-        )
-        i = (i + inc2) % mod
-        if i == end:
-            break
-
-    tmp += _to64((digest_c[31] << 8) | digest_c[30], 3)
-
-    return tmp
-
-
-def mysql_sha256_password_hash(password, salt):
-    """Return a MySQL compatible caching_sha2_password hash in raw format."""
-    if len(salt) != 20:
-        raise ValueError("Salt must be 20 characters long.")
-
-    count = 5
-    iteration = 1000 * count
-
-    digest = _sha256_digest(password, salt, iteration)
-    return "$A${0:>03}${1}{2}".format(count, salt, digest)
-
-
-def mysql_sha256_password_hash_hex(password, salt):
-    """Return a MySQL compatible caching_sha2_password hash in hex format."""
-    return mysql_sha256_password_hash(password, salt).encode().hex().upper()
-
-
 def _mysql_native_password(cleartext_password, salt=None):
     mysql_native_encrypted_password = "*" + sha1(sha1(to_bytes(cleartext_password)).digest()).hexdigest().upper()
     return mysql_native_encrypted_password
-
-
-def generate_random_salt(length=20):
-    salt = urlsafe_b64encode(urandom(length)).decode('utf-8')
-    return salt[:length]
 
 
 def _caching_sha2_password(cleartext_password, salt):
